@@ -1,16 +1,15 @@
 source 'http://rubygems.org'
 ruby '2.3.3'
 
-gem 'rails', '~> 4.2'
+gem 'rails', '= 5.0.0'
 
 gem 'citrus', '2.4.1'
 gem 'coffee-rails'
 gem 'haml-rails'
 gem 'mysql2'
-gem 'quiet_assets'
 gem 'rack'
 gem 'rack-cors', require: 'rack/cors'
-gem 'rake', '< 11.0'
+gem 'rake'
 gem 'request_store'
 gem 'sass-rails'
 gem 'sprockets-rails', require: 'sprockets/railtie'
@@ -18,7 +17,7 @@ gem 'uglifier'
 gem 'xml-simple'
 gem 'yui-compressor'
 
-gem 'activeadmin', '~> 1.0.0.pre2'
+gem 'activeadmin'
 gem 'acts_as_commentable_with_threading'
 gem 'acts_as_list'
 gem 'aws-sdk', '< 2.0'
@@ -32,7 +31,7 @@ gem 'diffy', require: false
 gem 'draper'
 gem 'ey_config'
 gem 'font-awesome-rails'
-gem 'foundation-rails'
+gem 'foundation-rails', '6.1.2.0'
 gem 'fuzzy-string-match', require: false
 gem 'gretel'
 gem 'high_voltage'
@@ -42,7 +41,7 @@ gem 'jquery-rails', '> 4.0'
 gem 'jquery-ui-rails', '> 5.0'
 gem 'newrelic_rpm'
 gem 'paper_trail', '< 8.0'
-gem 'paperclip'
+gem 'paperclip', '4.3.1'
 gem 'rails-observers'
 gem 'redcarpet'
 gem 'rouge'
@@ -50,7 +49,7 @@ gem 'ruby-progressbar'
 gem 'select2-rails'
 gem 'strip_attributes'
 gem 'sunspot_rails'
-gem 'sunspot_solr'
+gem 'sunspot_solr', '2.2.0'
 gem 'twitter-typeahead-rails'
 gem 'unread'
 gem 'will_paginate'
@@ -78,16 +77,8 @@ group :test do
   gem 'chromedriver-helper'
   gem 'cucumber-rails', require: false
   gem 'database_cleaner'
+  gem 'rails-controller-testing'
   gem 'selenium-webdriver', '>= 2.48' # works with firefox as of v34
   gem 'shoulda-matchers'
   gem 'webmock'
 end
-
-#####################################################
-# Profiling gems kept uncommented here because lazy.
-# See notes in `config/environments/production.rb`.
-# gem 'bullet'
-# gem 'rack-mini-profiler', require: false
-# gem 'flamegraph' # for rmp, optional
-# gem 'stackprof' # for rmp, optional
-#####################################################
