@@ -10,7 +10,7 @@
 #
 # See `ProtonymsWithReferencesMissingYear` for a database script for finding these.
 
-class Protonym < ActiveRecord::Base
+class Protonym < ApplicationRecord
   belongs_to :authorship, class_name: 'Citation', dependent: :destroy
   belongs_to :name
 
