@@ -1,17 +1,13 @@
-require File.expand_path '../boot', __FILE__
+require_relative 'boot'
 
 require 'rails/all'
 
 # Require the gems listed in Gemfile, including any gems
 # you've limited to :test, :development, or :production.
-Bundler.require *Rails.groups
+Bundler.require(*Rails.groups)
 
 module AntCat
   class Application < Rails::Application
-    # Settings in config/environments/* take precedence over those specified here.
-    # Application configuration should go into files in config/initializers
-    # -- all .rb files in that directory are automatically loaded.
-
     # Set Time.zone default to the specified zone and make Active Record auto-convert to this zone.
     # Run "rake -D time" for a list of tasks for finding time zone names. Default is UTC.
     # config.time_zone = 'Central Time (US & Canada)'
@@ -38,12 +34,6 @@ module AntCat
     config.autoload_paths += Dir["#{config.root}/app/services/**/"]
 
     config.action_dispatch.cookies_serializer = :hybrid
-
-    # Suppress deprecation warning.
-    # TODO remove after upgrading to Rails 5.
-    config.active_record.raise_in_transactional_callbacks = true
-
-    config.action_controller.action_on_unpermitted_parameters = :raise
 
     config.assets.enabled = true
   end
