@@ -1,4 +1,3 @@
-# Set up gems listed in the Gemfile.
-ENV['BUNDLE_GEMFILE'] ||= File.expand_path '../../Gemfile', __FILE__
+ENV['BUNDLE_GEMFILE'] ||= File.expand_path('../Gemfile', __dir__)
 
-require 'bundler/setup' if File.exist? ENV['BUNDLE_GEMFILE']
+require 'bundler/setup' # Set up gems listed in the Gemfile.
