@@ -1,1 +1,3 @@
-Rails.application.config.action_dispatch.cookies_serializer = :hybrid
+# Specify a serializer for the signed and encrypted cookie jars.
+# Valid options are :json, :marshal, and :hybrid.
+Rails.application.config.action_dispatch.cookies_serializer = :json
