@@ -1,7 +1,3 @@
-And(/^DEBUG$/) do
-  debugger
-end
-
 And(/^SHOT$/) do
   screenshot_and_save_page
 end
@@ -17,6 +13,10 @@ end
 
 And(/^WAIT$/) do
   sleep 1
+end
+
+And(/^WAIT_FOR_JQUERY$/) do
+  wait_for_jquery
 end
 
 Given(/^PENDING/) do
