@@ -44,17 +44,15 @@ def set_driver
   when "webkit"
     Capybara.javascript_driver = :webkit
   end
-
-  Capybara::Webkit.configure do |config|
-    config.block_unknown_urls
-  end
 end
+
+# TODO: Block unknown URLs.
 
 set_driver
 
 Capybara.default_max_wait_time = 5
 Capybara.default_selector = :css
-Capybara.save_and_open_page_path = './tmp/capybara'
+Capybara.save_path = './tmp/capybara'
 Capybara::Screenshot.prune_strategy = :keep_last_run
 
 ActionController::Base.allow_rescue = false
