@@ -11,7 +11,7 @@ describe Api::V1::CitationsController do
   describe "GET index" do
     # TODO depends on being run before any other specs.
     xit "gets all citations greater than a given number" do
-      get :index, starts_at: species.id # Get index starting at four.
+      get :index, params: { starts_at: species.id } # Get index starting at four.
 
       # Since we want no ids less than 4, we should get a starting id at 4.
       citations = JSON.parse response.body
