@@ -23,7 +23,7 @@ describe CatalogController do
       before { create :family }
 
       it "raises on taxon not found (=404 in prod)" do
-        expect { get :show, id: 99999 }.to raise_error ActiveRecord::RecordNotFound
+        expect { get :show, params: { id: 99999 }}.to raise_error ActiveRecord::RecordNotFound
       end
     end
   end
@@ -53,7 +53,7 @@ describe CatalogController do
 
     before do
       create_genus "Nylanderia"
-      get :autocomplete, q: "att", format: :json
+      get :autocomplete, params: { q: "att", format: :json }
     end
 
     it "returns matches" do
