@@ -21,14 +21,14 @@ describe ReferencesController do
         let!(:reference) { reference_factory author_name: 'E.O. Wilson', id: 99999 }
 
         it "redirects to #show" do
-          get :search, q: reference.id
+          get :search, params: { q: reference.id }
           expect(response).to redirect_to reference_path(reference)
         end
       end
 
       context "when reference does not exists" do
         it "does not redirect unless the reference exists" do
-          get :search, q: "11111"
+          get :search, params: { q: "11111" }
           expect(response).to render_template :search
         end
       end
@@ -44,7 +44,7 @@ describe ReferencesController do
       end
 
       it "autocompletes" do
-        get :autocomplete, q: "wilson", format: :json
+        get :autocomplete, params: { q: "wilson", format: :json }
 
         json = JSON.parse response.body
         expect(json.size).to eq 1
@@ -54,7 +54,7 @@ describe ReferencesController do
 
     context "when there are no matches" do
       it "returns an empty response" do
-        get :autocomplete, q: "willy", format: :json
+        get :autocomplete, params: { q: "willy", format: :json }
 
         json = JSON.parse response.body
         expect(json.size).to eq 0
@@ -69,7 +69,7 @@ describe ReferencesController do
         end
 
         it "autocompletes" do
-          get :autocomplete, q: "author:höll", format: :json
+          get :autocomplete, params: { q: "author:höll", format: :json }
 
           json = JSON.parse response.body
           expect(json.size).to eq 1
@@ -84,7 +84,7 @@ describe ReferencesController do
         end
 
         it "autocompletes" do
-          get :autocomplete, q: "author:abdul-ras", format: :json
+          get :autocomplete, params: { q: "author:abdul-ras", format: :json }
 
           json = JSON.parse response.body
           expect(json.size).to eq 1
