@@ -34,7 +34,7 @@ RSpec.configure do |config|
   end
 
   config.infer_spec_type_from_file_location!
-  config.include Devise::TestHelpers, type: :controller
+  config.include Devise::Test::ControllerHelpers, type: :controller
 
   # To avoid typing `FactoryBot.create` all the time (use `create`).
   config.include FactoryBot::Syntax::Methods
